part_two_blog_code/requirements.txt
ollama
python-dotenv
langfuse
gradio 